Add list of (id, amount) in Orders

+++ rapy/src/main/scala/models/User.scala
@@ -1,3 +1,3 @@
 package models
 
-case class User(val username: String, val locationId: Int, val balance: Int)
+case class User(val username: String, val locationId: Int, val balance: Float)
